fix travise by put the tests in one file #68
